Extract list.go to codeowners module
